GUI/requirements.txt
pyside6==6.11.0
pyserial==3.5